Refactor UI structure
- Replace key generation dialog with replacing the entire central widget
- Created tabs for key related information and smartcard related information
- Created 4 options in smartcard tab to edit name, public url, user pin and admin pin without dialog
- Refactored the widgets codes into various folders
