+++ tumpasrc/__init__.py
@@ -1,609 +1,85 @@
+import datetime
 import io
 import os
 import sys
 import time
-import datetime
-from PySide2 import QtWidgets
-from PySide2.QtCore import QObject, Signal, QSize, Qt, QThread
-from PySide2 import QtGui
 
 import johnnycanencrypt as jce
 import johnnycanencrypt.johnnycanencrypt as rjce
-from tumpasrc.resources import load_icon, load_css
+from PySide2 import QtGui, QtWidgets
+from PySide2.QtCore import QObject, QSize, Qt, QThread, Signal
+
+import tumpasrc.key_widgets.utils as key_utils
+from tumpasrc.commons import MessageDialogs, PasswordEdit, css
 from tumpasrc.configuration import get_keystore_directory
-
-css = load_css("mainwindow.css")
-
-
-class HardwareThread(QThread):
-    signal = Signal((bool,))
-
-    def __init__(self, nextsteps_slot):
-        QThread.__init__(self)
-        self.flag = True
-        self.signal.connect(nextsteps_slot)
-
-    def run(self):
-        while self.flag:
-            time.sleep(1)
-            result = rjce.is_smartcard_connected()
-            self.signal.emit(result)
+from tumpasrc.key_widgets.display import KeyWidgetList
+from tumpasrc.key_widgets.forms import NewKeyFormWidget
+from tumpasrc.resources import load_css, load_icon
+from tumpasrc.smartcard_widgets.forms import (SmartCardConfirmationDialog,
+                                              SmartCardTextFormWidget,
+                                              SmartPinFormWidget)
+from tumpasrc.threads import HardwareThread
 
 
-class PasswordEdit(QtWidgets.QLineEdit):
-    """
-    A LineEdit with icons to show/hide password entries
-    """
-
-    CSS = """QLineEdit {
-        border-radius: 10px;
-        height: 30px;
-        margin: 0px 0px 0px 0px;
-    }
-    """
-
-    def __init__(self):
-        super().__init__()
-
-        # Set styles
-        self.setStyleSheet(self.CSS)
-
-        self.visibleIcon = load_icon("eye_visible.svg")
-        self.hiddenIcon = load_icon("eye_hidden.svg")
-
-        self.setEchoMode(QtWidgets.QLineEdit.Password)
-        self.togglepasswordAction = self.addAction(
-            self.visibleIcon, QtWidgets.QLineEdit.TrailingPosition
+class KeyViewWidget(QtWidgets.QWidget):
+    def __init__(self, ks, parent=None):
+        super(KeyViewWidget, self).__init__(parent)
+        keyring_instruction_label = QtWidgets.QLabel(
+            "Single click on a key to enable writing to smart card. "
+            + "Double click on a key to export the public key."
         )
-        self.togglepasswordAction.triggered.connect(self.on_toggle_password_Action)
-        self.password_shown = False
+        keyring_instruction_label.setObjectName("keyring_instruction")
 
-    def on_toggle_password_Action(self):
-        if not self.password_shown:
-            self.setEchoMode(QtWidgets.QLineEdit.Normal)
-            self.password_shown = True
-            self.togglepasswordAction.setIcon(self.hiddenIcon)
-        else:
-            self.setEchoMode(QtWidgets.QLineEdit.Password)
-            self.password_shown = False
-            self.togglepasswordAction.setIcon(self.visibleIcon)
+        # List of keys shown
+        self.widget = KeyWidgetList(ks)
 
-
-class MessageDialogs:
-    """
-    A class that contains dialogue QMessageBoxes for success, error, etc.
-    """
-
-    @classmethod
-    def success_dialog(cls, msg: str):
-        success_dialog = QtWidgets.QMessageBox()
-        success_dialog.setText(f"{msg}")
-        success_dialog.setIcon(QtWidgets.QMessageBox.Information)
-        success_dialog.setWindowTitle("Success")
-        success_dialog.setStyleSheet(css)
-        return success_dialog
-
-    @classmethod
-    def error_dialog(cls, where: str, msg: str):
-        error_dialog = QtWidgets.QMessageBox()
-        error_dialog.setText(msg)
-        error_dialog.setIcon(QtWidgets.QMessageBox.Critical)
-        error_dialog.setWindowTitle(f"Error during {where}")
-        error_dialog.setStyleSheet(css)
-        return error_dialog
-
-
-class SmartCardConfirmationDialog(QtWidgets.QDialog):
-    # passphrase, adminpin
-    writetocard = Signal(
-        (str, str, int),
-    )
-
-    def __init__(
-        self,
-        nextsteps_slot,
-        title="Enter passphrase and pin for the smartcard",
-        firstinput="Key passphrase",
-        key=None,
-        enable_window=None,
-    ):
-        super(SmartCardConfirmationDialog, self).__init__()
-        self.setModal(True)
-        self.setFixedSize(600, 220)
-        self.setWindowTitle(title)
-        if enable_window:
-            self.rejected.connect(enable_window)
-        layout = QtWidgets.QFormLayout(self)
-        label = QtWidgets.QLabel(firstinput)
-        self.firstinput = firstinput
-        self.key = key
-        self.encryptionSubkey = QtWidgets.QCheckBox("Encryption")
-        self.encryptionSubkey.setEnabled(False)
-        self.signingSubkey = QtWidgets.QCheckBox("Signing")
-        self.signingSubkey.setEnabled(False)
-        self.authenticationSubkey = QtWidgets.QCheckBox("Authentication")
-        self.authenticationSubkey.setEnabled(False)
-        self.passphraseEdit = PasswordEdit()
-        layout.addRow(label, self.passphraseEdit)
-        label = QtWidgets.QLabel("Current Admin Pin")
-        self.addminPinEdit = PasswordEdit()
-        layout.addRow(label, self.addminPinEdit)
-        if self.key is not None:
-            label = QtWidgets.QLabel("Choose subkeys to upload:")
-            inhlayout = QtWidgets.QHBoxLayout()
-            got_enc, got_sign, got_auth = self.key.available_subkeys()
-            inhlayout.addWidget(self.encryptionSubkey)
-            inhlayout.addWidget(self.signingSubkey)
-            inhlayout.addWidget(self.authenticationSubkey)
-
-            if got_enc:
-                self.encryptionSubkey.setCheckState(Qt.Checked)
-                self.encryptionSubkey.setEnabled(True)
-            if got_sign:
-                self.signingSubkey.setCheckState(Qt.Checked)
-                self.signingSubkey.setEnabled(True)
-            if got_auth:
-                self.authenticationSubkey.setCheckState(Qt.Checked)
-                self.authenticationSubkey.setEnabled(True)
-            if any([got_enc, got_auth, got_sign]):  # Means we have at least one subkey
-                widget = QtWidgets.QWidget()
-                widget.setLayout(inhlayout)
-                # Now add in the formlayout
-                layout.addRow(label, widget)
-        widget = QtWidgets.QWidget()
-        widget.setLayout(layout)
-        # now the button
-        self.finalButton = QtWidgets.QPushButton(text="Write to smartcard")
-        self.finalButton.clicked.connect(self.getPassphrases)
-        vboxlayout = QtWidgets.QVBoxLayout()
-        vboxlayout.addWidget(widget)
-        vboxlayout.addWidget(self.finalButton)
-        self.setLayout(vboxlayout)
-        self.writetocard.connect(nextsteps_slot)
-        self.setStyleSheet(css)
-
-    def getPassphrases(self):
-        passphrase = self.passphraseEdit.text().strip()
-        adminpin = self.addminPinEdit.text().strip()
-        if len(adminpin) < 8:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details", "Admin pin must be 8 character or more."
-            )
-            self.error_dialog.show()
-            return
-        if len(passphrase) < 6:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details",
-                "{} must be 6 character or more.".format(self.firstinput),
-            )
-            self.error_dialog.show()
-            return
-
-        whichkeys = 0
-        if self.encryptionSubkey.checkState():
-            whichkeys += 1
-        if self.signingSubkey.checkState():
-            whichkeys += 2
-        if self.authenticationSubkey.checkState():
-            whichkeys += 4
-
-        # At least one subkey must be selected
-        if whichkeys == 0:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details", "At least one subkey must be selected"
-            )
-            self.error_dialog.show()
-            return
-
-        self.hide()
-
-        self.writetocard.emit(passphrase, adminpin, whichkeys)
-
-
-class SmartPinDialog(QtWidgets.QDialog):
-    # passphrase, adminpin
-    writetocard = Signal(
-        (str, str),
-    )
-
-    def __init__(
-        self,
-        nextsteps_slot,
-        title="Change user pin",
-        firstinput="New user pin",
-        enable_window=None,
-    ):
-        super(SmartPinDialog, self).__init__()
-        self.setModal(True)
-        self.setFixedSize(600, 220)
-        self.setWindowTitle(title)
-        if enable_window:
-            self.rejected.connect(enable_window)
-        layout = QtWidgets.QFormLayout(self)
-        label = QtWidgets.QLabel(firstinput)
-        self.firstinput = firstinput
-        self.passphraseEdit = PasswordEdit()
-        layout.addRow(label, self.passphraseEdit)
-        label = QtWidgets.QLabel("Current Admin Pin")
-        self.addminPinEdit = PasswordEdit()
-        layout.addRow(label, self.addminPinEdit)
-        widget = QtWidgets.QWidget()
-        widget.setLayout(layout)
-        # now the button
-        self.finalButton = QtWidgets.QPushButton(text="Write to smartcard")
-        self.finalButton.clicked.connect(self.getPassphrases)
-        vboxlayout = QtWidgets.QVBoxLayout()
-        vboxlayout.addWidget(widget)
-        vboxlayout.addWidget(self.finalButton)
-        self.setLayout(vboxlayout)
-        self.writetocard.connect(nextsteps_slot)
-        self.setStyleSheet(css)
-
-    def getPassphrases(self):
-        passphrase = self.passphraseEdit.text().strip()
-        adminpin = self.addminPinEdit.text().strip()
-        if len(adminpin) < 8:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details", "Admin pin must be 8 character or more."
-            )
-            self.error_dialog.show()
-            return
-        if self.firstinput == "New Admin pin" and len(passphrase) < 8:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details", "Admin pin must be 8 character or more."
-            )
-            self.error_dialog.show()
-            return
-        if len(passphrase) < 6:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details",
-                "{} must be 6 character or more.".format(self.firstinput),
-            )
-            self.error_dialog.show()
-            return
-
-        self.hide()
-
-        self.writetocard.emit(passphrase, adminpin)
-
-
-class SmartCardTextDialog(QtWidgets.QDialog):
-    # Public URL and Name
-    writetocard = Signal(
-        (str, str),
-    )
-
-    CSS = """QLineEdit {
-        border-radius: 10px;
-        height: 30px;
-        margin: 0px 0px 0px 0px;
-    }
-    """
-
-    def __init__(
-        self,
-        nextsteps_slot,
-        title="Enter public URL",
-        textInput="Public URL",
-        enable_window=None,
-    ):
-        super(SmartCardTextDialog, self).__init__()
-        self.setModal(True)
-        self.setFixedSize(600, 200)
-        self.setWindowTitle(title)
-        if enable_window:
-            self.rejected.connect(enable_window)
-        layout = QtWidgets.QFormLayout(self)
-        label = QtWidgets.QLabel(textInput)
-        self.textInput = textInput
-        self.textField = QtWidgets.QLineEdit("")
-        self.textField.setStyleSheet(self.CSS)
-        layout.addRow(label, self.textField)
-        label = QtWidgets.QLabel("Admin Pin")
-        self.adminPinEdit = PasswordEdit()
-        layout.addRow(label, self.adminPinEdit)
-        widget = QtWidgets.QWidget()
-        widget.setLayout(layout)
-        # now the button
-        self.finalButton = QtWidgets.QPushButton(text="Write to smartcard")
-        self.finalButton.clicked.connect(self.getTextValue)
-        vboxlayout = QtWidgets.QVBoxLayout()
-        vboxlayout.addWidget(widget)
-        vboxlayout.addWidget(self.finalButton)
-        self.setLayout(vboxlayout)
-        self.writetocard.connect(nextsteps_slot)
-        self.setStyleSheet(css)
-
-    def getTextValue(self):
-        text = self.textField.text().strip()
-        adminpin = self.adminPinEdit.text().strip()
-        if len(adminpin) < 8:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details", "Admin pin must be 8 character or more."
-            )
-            self.error_dialog.show()
-            return
-        if len(text) > 35:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details",
-                "{} must be less than 35 characters.".format(self.textInput),
-            )
-            self.error_dialog.show()
-            return
-        if not len(text):
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Editing smart card details",
-                "{} cannot be blank.".format(self.textInput),
-            )
-            self.error_dialog.show()
-            return
-
-        self.hide()
-        self.writetocard.emit(text, adminpin)
-
-
-class NewKeyDialog(QtWidgets.QDialog):
-    update_ui = Signal((jce.Key,))
-    disable_button = Signal()
-    enable_button = Signal()
-
-    def __init__(
-        self,
-        ks: jce.KeyStore,
-        newkey_slot,
-        disable_slot,
-        enable_slot,
-        enable_window=None,
-    ):
-        super(NewKeyDialog, self).__init__()
-        self.setModal(True)
-        self.update_ui.connect(newkey_slot)
-        self.disable_button.connect(disable_slot)
-        self.enable_button.connect(enable_slot)
-        self.ks = ks  # jce.KeyStore
-        self.setFixedSize(QSize(800, 600))
-        vboxlayout = QtWidgets.QVBoxLayout()
-        name_label = QtWidgets.QLabel("Your name:")
-        self.name_box = QtWidgets.QLineEdit("")
-        if enable_window:
-            self.rejected.connect(enable_window)
-
-        vboxlayout.addWidget(name_label)
-        vboxlayout.addWidget(self.name_box)
-
-        email_label = QtWidgets.QLabel("Email addresses (one email per line)")
-        self.email_box = QtWidgets.QPlainTextEdit()
-        self.email_box.setTabChangesFocus(True)
-
-        vboxlayout.addWidget(email_label)
-        vboxlayout.addWidget(self.email_box)
-        passphrase_label = QtWidgets.QLabel(
-            "Key Passphrase (recommended: 12+ chars in length):"
-        )
-        self.passphrase_box = PasswordEdit()
-
-        vboxlayout.addWidget(passphrase_label)
-        vboxlayout.addWidget(self.passphrase_box)
-
-        # now the checkboxes for subkey
-        self.encryptionSubkey = QtWidgets.QCheckBox("Encryption subkey")
-        self.encryptionSubkey.setCheckState(Qt.Checked)
-        self.signingSubkey = QtWidgets.QCheckBox("Signing subkey")
-        self.signingSubkey.setCheckState(Qt.Checked)
-        self.authenticationSubkey = QtWidgets.QCheckBox("Authentication subkey")
-
-        hboxlayout = QtWidgets.QHBoxLayout()
-        hboxlayout.addWidget(self.encryptionSubkey)
-        hboxlayout.addWidget(self.signingSubkey)
-        hboxlayout.addWidget(self.authenticationSubkey)
-
-        widget = QtWidgets.QWidget()
-        widget.setLayout(hboxlayout)
-        vboxlayout.addWidget(widget)
-
-        self.generateButton = QtWidgets.QPushButton("Generate")
-        self.generateButton.clicked.connect(self.generate)
-        self.generateButton.setMaximumWidth(50)
-        vboxlayout.addWidget(self.generateButton)
-
-        self.setLayout(vboxlayout)
-        self.setWindowTitle("Generate a new OpenPGP key")
-        self.setStyleSheet(css)
-
-    def generate(self):
-        self.generateButton.setEnabled(False)
-        emails = self.email_box.toPlainText()
-        name = self.name_box.text().strip()
-        password = self.passphrase_box.text().strip()
-
-        if not len(name):
-            self.error_dialog = MessageDialogs.error_dialog(
-                "generating new key", "Name cannot be blank."
-            )
-            self.error_dialog.show()
-            self.generateButton.setEnabled(True)
-            return
-
-        if not len(emails):
-            self.error_dialog = MessageDialogs.error_dialog(
-                "generating new key", "There must be at least one email."
-            )
-            self.error_dialog.show()
-            self.generateButton.setEnabled(True)
-            return
-
-        if not len(password):
-            self.error_dialog = MessageDialogs.error_dialog(
-                "generating new key", "Key passphrase cannot be blank."
-            )
-            self.error_dialog.show()
-            self.generateButton.setEnabled(True)
-            return
-
-        if len(password) < 6:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "generating new key",
-                "Key Passphrase must be at least 6 characters long.",
-            )
-            self.error_dialog.show()
-            self.generateButton.setEnabled(True)
-            return
-
-        # Now check which all subkeys are required
-        whichkeys = 0
-        if self.encryptionSubkey.checkState():
-            whichkeys += 1
-        if self.signingSubkey.checkState():
-            whichkeys += 2
-        if self.authenticationSubkey.checkState():
-            whichkeys += 4
-
-        # At least one subkey must be selected
-        if whichkeys == 0:
-            self.error_dialog = MessageDialogs.error_dialog(
-                "Generating new key", "At least one subkey must be selected"
-            )
-            self.error_dialog.show()
-            return
-
-        uids = []
-        for email in emails.split("\n"):
-            value = f"{name} <{email}>"
-            uids.append(value)
-        edate = datetime.datetime.now() + datetime.timedelta(days=3 * 365)
-        self.disable_button.emit()
-        # To make sure that the Generate button is disabled first
-        self.generateButton.setEnabled(False)
-        self.update()
-        self.repaint()
-        # Now let us try to create a key
-        newk = self.ks.create_newkey(
-            password,
-            uids,
-            ciphersuite=jce.Cipher.Cv25519,
-            expiration=edate,
-            subkeys_expiration=True,
-            whichkeys=whichkeys,
-        )
-        self.update_ui.emit(newk)
-        self.hide()
-        self.enable_button.emit()
-        self.success_dialog = MessageDialogs.success_dialog(
-            "Generated keys successfully!"
-        )
-        self.success_dialog.show()
-
-
-class KeyWidget(QtWidgets.QWidget):
-    SPACER = 14
-    BOTTOM_SPACER = 11
-
-    def __init__(self, key: jce.Key):
-        super(KeyWidget, self).__init__()
-        self.setObjectName("KeyWidgetItem")
-        self.setMinimumWidth(400)
-        self.setMinimumHeight(84)
-        self.key = key
-        fingerprint = key.fingerprint
-        self.fingerprint = fingerprint
-        self.keyfingerprint = QtWidgets.QLabel(fingerprint)
-        self.keyfingerprint.setObjectName("keyfingerprint")
-        date = key.creationtime.date()
-        date_label = QtWidgets.QLabel(f"Created at: {date.strftime('%Y-%m-%d')}")
-        date_label.setAlignment(Qt.AlignTop)
-        date_label.setContentsMargins(0, 0, 0, 0)
-
-        # UIDs
-        uid_vboxlayout = QtWidgets.QVBoxLayout()
-        uid_vboxlayout.setSpacing(0)
-        uid_vboxlayout.setContentsMargins(0, 0, 0, 0)
-        for uid in key.uids:
-            uid_label = QtWidgets.QLabel(uid["value"])
-            uid_vboxlayout.addWidget(uid_label)
-        uid_widget = QtWidgets.QWidget()
-        uid_widget.setLayout(uid_vboxlayout)
-
-        # UID and date layout
+        # Buttons
+        self.generateButton = QtWidgets.QPushButton(text="Generate new key")
+        self.generateButton.clicked.connect(self.parent().show_generate_dialog)
+        self.uploadButton = QtWidgets.QPushButton(text="Upload to SmartCard")
+        self.uploadButton.clicked.connect(self.parent().upload_to_smartcard)
+        self.uploadButton.setEnabled(False)
         hlayout = QtWidgets.QHBoxLayout()
-        hlayout.addWidget(uid_widget)
-        hlayout.addWidget(date_label)
-        hlayout.setAlignment(Qt.AlignTop)
-        hlayout.setContentsMargins(11, 0, 11, 11)
-        group_widget = QtWidgets.QWidget()
-        group_widget.setLayout(hlayout)
+        hlayout.addWidget(self.generateButton)
+        hlayout.addWidget(self.uploadButton)
+        wd = QtWidgets.QWidget()
+        wd.setLayout(hlayout)
 
-        fp_group_layout = QtWidgets.QVBoxLayout()
-        fp_group_layout.addWidget(self.keyfingerprint)
-        fp_group_layout.addWidget(group_widget)
-
-        self.setLayout(fp_group_layout)
-        self.setToolTip("Double click to export public key")
-        self.setObjectName("keywidget")
-
-    def mouseDoubleClickEvent(self, event):
-        if self.export_public_key(self, self.fingerprint, self.key.get_pub_key()):
-            self.success_dialog = MessageDialogs.success_dialog(
-                "Exported public key successfully!"
-            )
-            self.success_dialog.show()
-
-    @classmethod
-    def export_public_key(cls, widget, fingerprint, public_key):
-        select_path = QtWidgets.QFileDialog.getExistingDirectory(
-            widget,
-            "Select directory to save public key",
-            ".",
-            QtWidgets.QFileDialog.ShowDirsOnly,
-        )
-        if select_path:
-            filepassphrase = f"{fingerprint}.pub"
-            filepath = os.path.join(select_path, filepassphrase)
-            with open(filepath, "w") as fobj:
-                fobj.write(public_key)
-            return True
-        return False
+        vboxlayout = QtWidgets.QVBoxLayout()
+        vboxlayout.addWidget(keyring_instruction_label)
+        vboxlayout.addWidget(self.widget)
+        vboxlayout.addWidget(wd)
+        self.setLayout(vboxlayout)
 
 
-class KeyWidgetList(QtWidgets.QListWidget):
-    def __init__(self, ks):
-        super(KeyWidgetList, self).__init__()
-        self.setObjectName("KeyWidgetList")
-        self.ks = ks
+class SmartCardViewWidget(QtWidgets.QWidget):
+    def __init__(self, parent=None):
+        super(SmartCardViewWidget, self).__init__(parent)
+        self.main_area = QtWidgets.QLabel("This is where smart card stuff go.")
 
-        # Set layout.
-        # self.layout = QtWidgets.QVBoxLayout(self)
-        # self.setLayout(self.layout)
-        self.updateList()
-        self.setSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Minimum)
-        self.setMinimumHeight(400)
-        self.currentItemChanged.connect(self.on_item_changed)
+        # Left Navbar
+        self.editNameButton = QtWidgets.QPushButton(text="Edit Name")
+        self.editNameButton.clicked.connect(self.parent().show_set_name)
+        self.editPublicUrlButton = QtWidgets.QPushButton(text="Edit Public URL")
+        self.editPublicUrlButton.clicked.connect(self.parent().show_set_public_url)
+        self.editUserPinButton = QtWidgets.QPushButton(text="Edit User Pin")
+        self.editUserPinButton.clicked.connect(self.parent().show_change_user_pin)
+        self.editAdminPinButton = QtWidgets.QPushButton(text="Edit Admin Pin")
+        self.editAdminPinButton.clicked.connect(self.parent().show_change_admin_pin)
+        vnavlayout = QtWidgets.QVBoxLayout()
+        vnavlayout.addWidget(self.editNameButton)
+        vnavlayout.addWidget(self.editPublicUrlButton)
+        vnavlayout.addWidget(self.editUserPinButton)
+        vnavlayout.addWidget(self.editAdminPinButton)
+        navbarWidget = QtWidgets.QWidget()
+        navbarWidget.setLayout(vnavlayout)
+        navbarWidget.setObjectName("sidenavbar")
+        navbarWidget.setMaximumWidth(180)
 
-    def updateList(self):
-        try:
-            keys = self.ks.get_all_keys()
-            keys.sort(key=lambda x: x.creationtime, reverse=True)
-            for key in keys:
-                kw = KeyWidget(key)
-                item = QtWidgets.QListWidgetItem()
-                item.setSizeHint(kw.sizeHint())
-                self.addItem(item)
-                self.setItemWidget(item, kw)
-            if len(keys) > 0:
-                # Select the top most row
-                self.setCurrentRow(0)
-        except Exception as e:
-            print(e)
-
-    def on_item_changed(self):
-        print(self.selectedItems())
-
-    def addnewKey(self, key):
-        kw = KeyWidget(key)
-        item = QtWidgets.QListWidgetItem()
-        item.setSizeHint(kw.sizeHint())
-        self.insertItem(0, item)
-        self.setItemWidget(item, kw)
-        self.setCurrentRow(0)
+        self.hlayout = QtWidgets.QHBoxLayout()
+        self.hlayout.addWidget(navbarWidget)
+        self.hlayout.addWidget(self.main_area)
+        self.setLayout(self.hlayout)
 
 
 class MainWindow(QtWidgets.QMainWindow):
@@ -615,8 +91,6 @@ class MainWindow(QtWidgets.QMainWindow):
         self.setMaximumWidth(600)
         self.setMaximumHeight(575)
         self.ks = jce.KeyStore(get_keystore_directory())
-        self.vboxlayout_for_keys = QtWidgets.QVBoxLayout()
-        self.widget = KeyWidgetList(self.ks)
         self.current_fingerprint = ""
         self.cardcheck_thread = HardwareThread(self.enable_upload)
 
@@ -632,9 +106,9 @@ class MainWindow(QtWidgets.QMainWindow):
 
         # smartcard menu
         changepinAction = QtWidgets.QAction("Change user &pin", self)
-        changepinAction.triggered.connect(self.show_change_user_pin_dialog)
+        changepinAction.triggered.connect(self.show_change_user_pin)
         changeadminpinAction = QtWidgets.QAction("Change &admin pin", self)
-        changeadminpinAction.triggered.connect(self.show_change_admin_pin_dialog)
+        changeadminpinAction.triggered.connect(self.show_change_admin_pin)
         changenameAction = QtWidgets.QAction("Set cardholder &name", self)
         changenameAction.triggered.connect(self.show_set_name)
         changeurlAction = QtWidgets.QAction("Set public key &URL", self)
@@ -648,34 +122,13 @@ class MainWindow(QtWidgets.QMainWindow):
         smartcardmenu.addAction(changeurlAction)
         smartcardmenu.addAction(resetYubiKeylAction)
 
-        self.cwidget = QtWidgets.QWidget()
-        self.generateButton = QtWidgets.QPushButton(text="Generate new key")
-        self.generateButton.clicked.connect(self.show_generate_dialog)
-        self.uploadButton = QtWidgets.QPushButton(text="Upload to SmartCard")
-        self.uploadButton.clicked.connect(self.upload_to_smartcard)
-        self.uploadButton.setEnabled(False)
         # self.widget.itemSelectionChanged.connect(self.enable_upload)
-
-        hlayout = QtWidgets.QHBoxLayout()
-        hlayout.addWidget(self.generateButton)
-        hlayout.addWidget(self.uploadButton)
-        wd = QtWidgets.QWidget()
-        wd.setLayout(hlayout)
-
-        keyring_label = QtWidgets.QLabel("Available keys")
-        keyring_label.setObjectName("keyring_label")
-        keyring_instruction_label = QtWidgets.QLabel(
-            "Single click on a key to enable writing to smart card. "
-            + "Double click on a key to export the public key."
-        )
-        keyring_instruction_label.setObjectName("keyring_instruction")
-        vboxlayout = QtWidgets.QVBoxLayout()
-        vboxlayout.addWidget(keyring_label)
-        vboxlayout.addWidget(keyring_instruction_label)
-        vboxlayout.addWidget(self.widget)
-        vboxlayout.addWidget(wd)
-        self.cwidget.setLayout(vboxlayout)
-        self.setCentralWidget(self.cwidget)
+        self.tabs = QtWidgets.QTabWidget()
+        self.keyWidget = KeyViewWidget(self.ks, self)
+        self.tabs.addTab(self.keyWidget, "Available Keys")
+        self.smartCardWidget = SmartCardViewWidget(self)
+        self.tabs.addTab(self.smartCardWidget, "Smart Card Settings")
+        self.setCentralWidget(self.tabs)
         self.setStyleSheet(css)
         self.cardcheck_thread.start()
 
@@ -707,53 +160,64 @@ class MainWindow(QtWidgets.QMainWindow):
         "Slot to enable the upload to smartcard button"
         # If no item is selected on the ListWidget, then
         # no need to update the uploadButton status.
-        if not self.widget.selectedItems():
+        try:
+            if not self.keyWidget.widget.selectedItems():
+                return
+        except RuntimeError:
             return
-        self.uploadButton.setEnabled(value)
+        self.keyWidget.uploadButton.setEnabled(value)
 
-    def show_change_user_pin_dialog(self):
-        "This slot shows the input dialog to change user pin"
+    def show_change_user_pin(self):
+        "This slot shows the input widget to change user pin"
         self.cardcheck_thread.flag = False
-        self.smalldialog = SmartPinDialog(
+        self.tabs.setCurrentIndex(1)
+        self.smartCardWidget.hlayout.removeWidget(self.smartCardWidget.main_area)
+        self.smartCardWidget.main_area.deleteLater()
+        self.smartCardWidget.main_area = SmartPinFormWidget(
             self.change_pin_on_card_slot,
             "Change user pin",
             "New User pin",
-            enable_window=self.enable_mainwindow,
         )
-        self.smalldialog.show()
+        self.smartCardWidget.hlayout.addWidget(self.smartCardWidget.main_area)
 
     def show_set_public_url(self):
-        "This slot shows the input dialog to set public url"
+        "This slot shows the input widget to set public url"
         self.cardcheck_thread.flag = False
-        self.smalldialog = SmartCardTextDialog(
+        self.tabs.setCurrentIndex(1)
+        self.smartCardWidget.hlayout.removeWidget(self.smartCardWidget.main_area)
+        self.smartCardWidget.main_area.deleteLater()
+        self.smartCardWidget.main_area = SmartCardTextFormWidget(
             self.set_url_on_card_slot,
             "Add public URL",
             "Public URL",
-            enable_window=self.enable_mainwindow,
         )
-        self.smalldialog.show()
+        self.smartCardWidget.hlayout.addWidget(self.smartCardWidget.main_area)
 
     def show_set_name(self):
-        "This slot shows the input dialog to set name"
+        "This slot shows the input widget to set name"
         self.cardcheck_thread.flag = False
-        self.smalldialog = SmartCardTextDialog(
+        self.tabs.setCurrentIndex(1)
+        self.smartCardWidget.hlayout.removeWidget(self.smartCardWidget.main_area)
+        self.smartCardWidget.main_area.deleteLater()
+        self.smartCardWidget.main_area = SmartCardTextFormWidget(
             self.set_name_on_card_slot,
             "Add Name",
             "Name",
-            enable_window=self.enable_mainwindow,
         )
-        self.smalldialog.show()
+        self.smartCardWidget.hlayout.addWidget(self.smartCardWidget.main_area)
 
-    def show_change_admin_pin_dialog(self):
-        "This slot shows the input dialog to change admin pin"
+    def show_change_admin_pin(self):
+        "This slot shows the input widget to change admin pin"
         self.cardcheck_thread.flag = False
-        self.smalldialog = SmartPinDialog(
+        self.tabs.setCurrentIndex(1)
+        self.smartCardWidget.hlayout.removeWidget(self.smartCardWidget.main_area)
+        self.smartCardWidget.main_area.deleteLater()
+        self.smartCardWidget.main_area = SmartPinFormWidget(
             self.change_admin_pin_on_card_slot,
             "Change admin pin",
             "New Admin pin",
-            enable_window=self.enable_mainwindow,
         )
-        self.smalldialog.show()
+        self.smartCardWidget.hlayout.addWidget(self.smartCardWidget.main_area)
 
     def change_pin_on_card_slot(self, userpin, adminpin):
         "Final slot which will try to change the userpin"
@@ -818,19 +282,6 @@ class MainWindow(QtWidgets.QMainWindow):
         self.success_dialog.show()
         self.enable_cardcheck_thread_slot()
 
-    def show_generate_dialog(self):
-        "Shows the dialog to generate new key"
-        self.disable_cardcheck_thread_slot()
-        self.newd = NewKeyDialog(
-            self.ks,
-            self.widget.addnewKey,
-            self.disable_generate_button,
-            self.enable_generate_button,
-            enable_window=self.enable_mainwindow,
-        )
-        self.newd.show()
-        self.setEnabled(False)
-
     def disable_generate_button(self):
         self.cardcheck_thread.flag = False
         self.generateButton.setEnabled(False)
@@ -844,12 +295,28 @@ class MainWindow(QtWidgets.QMainWindow):
         self.update()
         self.repaint()
 
+    def show_generate_dialog(self):
+        "Shows the form to generate new key"
+        self.disable_cardcheck_thread_slot()
+        new_key_form = NewKeyFormWidget(self.ks, self.restore_list_view)
+        self.setCentralWidget(new_key_form)
+
+    def restore_list_view(self):
+        self.enable_cardcheck_thread_slot()
+        self.tabs = QtWidgets.QTabWidget()
+        self.keyWidget = KeyViewWidget(self.ks, self)
+        self.tabs.addTab(self.keyWidget, "Available Keys")
+        self.smartCardWidget = SmartCardViewWidget(self)
+        self.tabs.addTab(self.smartCardWidget, "Smart Card Settings")
+        self.setCentralWidget(self.tabs)
+
     def enable_mainwindow(self):
         self.enable_cardcheck_thread_slot()
         self.setEnabled(True)
 
     def disable_cardcheck_thread_slot(self):
         self.cardcheck_thread.flag = False
+        self.cardcheck_thread.quit()
 
     def enable_cardcheck_thread_slot(self):
         self.cardcheck_thread.flag = True
@@ -858,7 +325,7 @@ class MainWindow(QtWidgets.QMainWindow):
     def upload_to_smartcard(self):
         "Shows the userinput dialog to upload the selected key to the smartcard"
         # This means no key is selected on the list
-        if not self.widget.selectedItems():
+        if not self.keyWidget.widget.selectedItems():
             self.error_dialog = MessageDialogs.error_dialog(
                 "upload to smart card", "Please select a key from the list."
             )
@@ -867,8 +334,8 @@ class MainWindow(QtWidgets.QMainWindow):
 
         self.disable_cardcheck_thread_slot()
         self.setEnabled(False)
-        item = self.widget.selectedItems()[0]
-        kw = self.widget.itemWidget(item)
+        item = self.keyWidget.widget.selectedItems()[0]
+        kw = self.keyWidget.widget.itemWidget(item)
         self.current_key = kw.key
         self.sccd = SmartCardConfirmationDialog(
             self.get_pins_and_passphrase_and_write,
@@ -901,16 +368,16 @@ class MainWindow(QtWidgets.QMainWindow):
 
     def export_public_key(self):
         # This means no key is selected on the list
-        if not self.widget.selectedItems():
+        if not self.keyWidget.widget.selectedItems():
             self.error_dialog = MessageDialogs.error_dialog(
                 "exporting public key", "Please select a key from the list."
             )
             self.error_dialog.show()
             return
 
-        item = self.widget.selectedItems()[0]
-        kw = self.widget.itemWidget(item)
-        if KeyWidget.export_public_key(self, kw.key.fingerprint, kw.key.get_pub_key()):
+        item = self.keyWidget.widget.selectedItems()[0]
+        kw = self.keyWidget.widget.itemWidget(item)
+        if key_utils.export_public_key(self, kw.key.fingerprint, kw.key.get_pub_key()):
             self.success_dialog = MessageDialogs.success_dialog(
                 "Exported public key successfully!"
             )
