+++ tumpasrc/resources/css/mainwindow.css
@@ -22,7 +22,8 @@ QLineEdit {
     height: 40px;
     margin: 0px 0px 0px 0px;
     padding-left: 5px;
-    border-radius: 10px;
+    border-radius: 5px;
+    border: 1px solid black;
 }
 
 QLabel#keyring_label {
@@ -34,12 +35,23 @@ QLabel#keyring_instruction {
 }
 
 QPlainTextEdit {
-    border-radius: 20px;
+    border-radius: 5px;
     background-color: palette(base);
     padding-left: 5px;
     padding-top: 5px;
+    border: 1px solid black;
 }
 
+QWidget#sidenavbar {
+    background-color: #D1E5F4;
+}
+QWidget#sidenavbar QPushButton {
+    border-color: #2D7AD4;
+    font-size: 14px;
+    background-color: #D1E5F4;
+    color: black;
+    min-width: 0;
+}
 
 QLabel#keyfingerprint {
     font-size: 18px;
